Add missing cast to float in dB->linear calculation

+++ src/SoundPlayer.cpp
@@ -523,7 +523,7 @@ ZunResult SoundPlayer::InitSoundBuffers()
         for (idx = 0; idx < ARRAY_SIZE(g_SoundBufferIdxVol); idx++)
         {
             if (this->LoadSound(idx, g_SFXList[g_SoundBufferIdxVol[idx].bufferIdx], 
-                                1.0f / std::powf(10.0f, g_SoundBufferIdxVol[idx].volume / -2000)) != ZUN_SUCCESS)
+                                1.0f / std::powf(10.0f, (float)g_SoundBufferIdxVol[idx].volume / -2000)) != ZUN_SUCCESS)
             {
                 GameErrorContext::Log(&g_GameErrorContext, TH_ERR_SOUNDPLAYER_FAILED_TO_LOAD_SOUND_FILE,
                                       g_SFXList[idx]);
